backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
chromadb
pymupdf
requests
python-multipart 